search page component reorder
