+++ client/src/components/atoms/Form/Input.module.scss
@@ -22,7 +22,7 @@
     composes: inputWrap;
 
     .input {
-        padding-left: $spacer * 1.5;
+        padding-left: $spacer;
     }
 
     svg {
