--- client/src/routes/Search/Filters.module.scss
@@ -1,16 +1,0 @@
-@import '../../styles/variables';
-
-.filter {
-    ul {
-        padding-left: 0;
-    }
-}
-
-.filterTitle {
-    font-size: $font-size-base;
-    color: $brand-grey-light;
-    margin-top: 0;
-    border-bottom: 1px solid $brand-grey-lighter;
-    padding-bottom: $spacer / 4;
-    margin-bottom: $spacer / 4;
-}
