+++ client/src/routes/Search/Sidebar.module.scss
@@ -11,3 +11,18 @@
         }
     }
 }
+
+.filter {
+    ul {
+        padding-left: 0;
+    }
+}
+
+.filterTitle {
+    font-size: $font-size-base;
+    color: $brand-grey-light;
+    margin-top: 0;
+    border-bottom: 1px solid $brand-grey-lighter;
+    padding-bottom: $spacer / 4;
+    margin-bottom: $spacer / 4;
+}
