+++ client/src/routes/Search/index.module.scss
@@ -4,8 +4,9 @@
     display: grid;
     grid-gap: $spacer * 2;
     grid-template-columns: 1fr;
+    margin-top: $spacer * 2;
 
     @media (min-width: $break-point--medium) {
-        grid-template-columns: 1fr 3fr;
+        grid-template-columns: 3fr 1fr;
     }
 }
